backend/check_models.py
import google.generativeai as genai
import os
from dotenv import load_dotenv

# Load the API key from your .env file
load_dotenv()
try:
    genai.configure(api_key=os.getenv("GOOGLE_API_KEY"))
except Exception as e:
    print(f"Failed to configure API: {e}")
    exit()

print("\n--- Checking for available models for your API key ---")

# List the available models
for m in genai.list_models():
  # Check if the model supports the 'generateContent' method
  if 'generateContent' in m.supported_generation_methods:
    print(f"Model found: {m.name}")

print("\n--- Check complete ---")